backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "authlib>=1.6.8",
    "bcrypt>=4.0.0",
    "email-validator>=2.0.0",
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.1.0",
    "langchain>=1.2.10",
    "openai>=2.24.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.8.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn[standard]>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
